+++ handlebars/src/test/scala/io/fintrospect/templating/HandlebarsTemplatesTest.scala
@@ -1,0 +1,3 @@
+package io.fintrospect.templating
+
+class HandlebarsTemplatesTest extends TemplatesSpec(HandlebarsTemplates, "handlebars")