+++ handlebars/src/test/scala/io/fintrospect/templating/RenderHandlebarsViewTest.scala
@@ -1,8 +1,0 @@
-package io.fintrospect.templating
-
-import com.gilt.handlebars.scala.Handlebars
-import io.fintrospect.formats.Html
-
-class RenderHandlebarsViewTest extends RenderTemplateEngineSpec[Handlebars[Any]](HandlebarsTemplateLoaders, "handlebars") {
-  override def renderViewFor(loader: TemplateLoader[Handlebars[Any]]) = new RenderHandlebarsView(Html.ResponseBuilder, loader)
-}